requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
scikit-learn
Pillow
tqdm
